Move limelight helpers to utilities
